Fix Operator Resolution
comp: Fixed an issue that caused operator resolution for ternary
operators to fail in some cases.

+++ src/compiler/dyvil/tools/compiler/ast/context/CombiningContext.java
@@ -110,7 +110,7 @@ public class CombiningContext implements IContext
 	}
 
 	@Override
-	public IOperator resolveOperator(Name name, int type)
+	public IOperator resolveOperator(Name name, byte type)
 	{
 		final IOperator inner = this.inner.resolveOperator(name, type);
 		if (inner == null || inner.getType() != type)
